functions/requirements.txt
azure-functions>=1.19
lightgbm>=4.0
numpy>=1.24
onnxruntime>=1.16
onnxmltools>=1.11
pandas>=2.0
psycopg[binary]>=3.1
requests>=2.31
scikit-learn>=1.3
skl2onnx>=1.16
